+++ crashtracker/src/crash_info/structured_log_format/metadata.rs
@@ -1,0 +1,16 @@
+// Copyright 2024-Present Datadog, Inc. https://www.datadoghq.com/
+// SPDX-License-Identifier: Apache-2.0
+
+use chrono::{DateTime, Utc};
+use serde::{Deserialize, Serialize};
+use std::collections::HashMap;
+use uuid::Uuid;
+
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+pub struct Metadata {
+    pub library_name: String,
+    pub library_version: String,
+    pub family: String,
+    // Should include "service", "environment", etc
+    pub tags: Vec<String>,
+}
